Add new DTOs, service, mappers and controller in accounts module
This commit introduces a collection of new Data Transfer Objects (DTOs) these are AccountsDto, CustomerDto, ResponseDto, and ErrorResponseDto. In addition to these DTOs, an AccountService, two mappers AccountMapper, and CustomerMapper - and an AccountsController are also added. A new dataSource.xml file is also included in the commit. The changes contribute to creating, managing, and mapping account-related data.