Add customer account deletion functionality
Two modifications have been done: First, a new method `deleteByCustomerId` is added in the AccountRepository which allows the deletion of an account by the customer ID. This method is annotated with `@Modifying` and `@Transactional` to indicate it modifies the database. Secondly, an endpoint (`/delete`) has been added in AccountsController to handle account deletion requests based on a provided mobile number. The `deleteAccountDetails` service employs the new repository method to achieve this.