Switch to jib-maven plugin and add build version endpoint
The commit includes changing the build process from the spring-boot-maven-plugin to the jib-maven-plugin within the pom.xml files for accounts, loans, and cards. A Dockerfile for the account service was also removed. Furthermore, a "build-info" endpoint was added to the AccountsController to display the build version, which was added to the application.yml file.