Add contact information to Accounts microservice
A new feature is introduced in the Accounts microservice to return contact information. A new data transfer object (DTO), AccountContactInfoDto, has been implemented. It is also added to AccountsApplication to allow it to be configurable. This includes adding a new API endpoint, 'contact-info', in AccountsController to access this information. Relevant properties are updated in application.yml.