Update config setting names and identifiers
The configuration settings in the YAML files and DTOs have been updated. Changes include modifications of the git URI to HTTPS, correction of readiness and liveness state keys, and changes in configuration property prefixes for card and loans services. Furthermore, rearranges the port configuration to the end of the file and adjusts the active profile setting in cards application.