Add new 'accounts' microservice and update .gitignore files
This commit introduces a new 'accounts' microservice, complete with models, repositories, and controller logic. It also includes updates to the '.gitignore' files for respective microservices to ensure certain files and directories are not tracked by git. The 'accounts' microservice enables the creation, deletion, and updating of account information, and communicates with Eureka server and pulls configuration from the 'configserver'.