Add fallback controller and circuit breaker for service routing
A new fallback controller has been added to handle errors if the service routing fails. This controller returns a mono where it sends a message for the user to try after some time or contact the support team. Additionally, a circuit breaker has been configured to the /eazybank/accounts route, which falls back to the FallbackController in case of failure. The parameters for the circuit breaker have also been added in the application.yml file.