Add Redis service and update docker images
This update adds a redis service to the docker compose configuration file for QA, default, and prod environments. It also updates the versions for Configserver, EurekaServer, Accounts, Loans, Cards, and GatewayServer services from s9 to s10. The addition of the Redis service will help to enhance the capabilities of the application.